feat: create login, auth module and protect store routes to only admi…

+++ src/stores/controllers/stores.controller.ts
@@ -6,13 +6,20 @@ import {
   Param,
   Delete,
   Query,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { StoresService } from '../services/stores.service';
 import { StoreDtoRequest } from '../dtos/requests/store-request.dto';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 import { StoreDtoResponse } from '../dtos/responses/store-response.dto';
+import { AuthGuard } from 'src/guards/auth.guard';
+import { IsCustomer } from 'src/decorators/customer.decorator';
+import { CustomerTypeEnum } from 'src/customers/common/enums/customer-type.enum';
+import { CustomerDtoResponse } from 'src/customers/dto/responses/customer-response.dto';
 
 @ApiTags('Stores')
+@UseGuards(AuthGuard)
 @Controller('api/v1/stores')
 export class StoresController {
   constructor(private readonly storesService: StoresService) {}
@@ -24,8 +31,15 @@ export class StoresController {
     type: StoreDtoResponse,
   })
   @ApiResponse({ status: 400, description: 'Invalid request' })
-  create(@Body() dto: StoreDtoRequest): Promise<StoreDtoResponse> {
-    return this.storesService.create(dto);
+  create(
+    @IsCustomer() customer: CustomerDtoResponse,
+    @Body() dto: StoreDtoRequest,
+  ): Promise<StoreDtoResponse> {
+    if (customer.customerType !== CustomerTypeEnum.ADMIN) {
+      throw new ForbiddenException('Access denied.');
+    }
+
+    return this.storesService.create(customer.id, dto);
   }
 
   @Get('by-user')
@@ -35,7 +49,12 @@ export class StoresController {
     type: StoreDtoResponse,
   })
   @ApiResponse({ status: 404, description: 'Store not found.' })
-  findAllByUser(): Promise<StoreDtoResponse[]> {
+  findAllByUser(
+    @IsCustomer() customer: CustomerDtoResponse,
+  ): Promise<StoreDtoResponse[]> {
+    if (customer.customerType !== CustomerTypeEnum.ADMIN) {
+      throw new ForbiddenException('Access denied.');
+    }
     return this.storesService.findAllByUser();
   }
 
@@ -46,14 +65,26 @@ export class StoresController {
     type: StoreDtoResponse,
   })
   @ApiResponse({ status: 404, description: 'Store not found.' })
-  searchById(@Query('storeId') storeId: string): Promise<StoreDtoResponse> {
+  searchById(
+    @IsCustomer() customer: CustomerDtoResponse,
+    @Query('storeId') storeId: string,
+  ): Promise<StoreDtoResponse> {
+    if (customer.customerType !== CustomerTypeEnum.ADMIN) {
+      throw new ForbiddenException('Access denied.');
+    }
     return this.storesService.searchById(+storeId);
   }
 
   @Delete(':storeId')
   @ApiResponse({ status: 204, description: 'Remove store' })
   @ApiResponse({ status: 404, description: 'Store not found.' })
-  remove(@Param('storeId') storeId: string) {
+  remove(
+    @IsCustomer() customer: CustomerDtoResponse,
+    @Param('storeId') storeId: string,
+  ) {
+    if (customer.customerType !== CustomerTypeEnum.ADMIN) {
+      throw new ForbiddenException('Access denied.');
+    }
     return this.storesService.remove(+storeId);
   }
 }
